requirements.txt
streamlit
pandas
numpy
yfinance
scikit-learn
tensorflow
requests
textblob
openai
python-dotenv